separated out the styles sheet by restructuring the folders
