+++ client/src/components/index.tsx
@@ -1,1 +1,1 @@
-export { default as Navigation } from "./Navigation";
+export { default as Navigation } from "./Navigation/Navigation";
