+++ client/src/pages/Login/Login.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Container } from "./styles";
+
+const Login = () => {
+  return (
+    <Container>
+      <h1>Login Page</h1>
+    </Container>
+  );
+};
+
+export default Login;
