+++ client/src/pages/Login/styles.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Container = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: calc(100vh - 70px);
+`;
