--- client/src/pages/Signup.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Signup = () => {
-  return <div>Signup Page</div>;
-};
-
-export default Signup;
