+++ client/src/pages/Signup/Signup.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { Container } from "./styles";
+
+const Signup = () => {
+  return (
+    <Container>
+      <h1>Signup Page</h1>
+    </Container>
+  );
+};
+
+export default Signup;
