+++ client/src/pages/index.tsx
@@ -1,3 +1,3 @@
-export { default as Welcome } from "./Welcome";
-export { default as Login } from "./Login";
-export { default as Signup } from "./Signup";
+export { default as Welcome } from "./Welcome/Welcome";
+export { default as Login } from "./Login/Login";
+export { default as Signup } from "./Signup/Signup";
